move rawSchema to pacakge object
