+++ s2ml/src/main/scala/com/kakao/ml/io/package.scala
@@ -1,6 +1,7 @@
 package com.kakao.ml
 
 import org.apache.spark.sql.ColumnName
+import org.apache.spark.sql.types.{StringType, LongType, StructField, StructType}
 
 package object io {
 
@@ -22,5 +23,16 @@ package object io {
   val indexedUserCol = new ColumnName(indexedUserColString)
   val indexedItemCol = new ColumnName(indexedItemColString)
   val confidenceCol = new ColumnName(confidenceColString)
+
+  val rawSchema = StructType(Seq(
+    StructField("log_ts", LongType),
+    StructField("operation", StringType),
+    StructField("log_type", StringType),
+    StructField("edge_from", StringType),
+    StructField("edge_to", StringType),
+    StructField("label", StringType),
+    StructField("props", StringType),
+    StructField("service", StringType),
+    StructField("date_id", StringType)))
 
 }
